UPD move example to examples dir
